+++ 8vim/src/main/java/inc/flide/vim8/views/NumberKeypadView.java
@@ -1,4 +1,4 @@
-package inc.flide.vim8.ui.views;
+package inc.flide.vim8.views;
 
 import android.content.Context;
 import com.hijamoya.keyboardview.Keyboard;
