+++ 8vim/src/main/java/inc/flide/vim8/views/mainkeyboard/MainKeyboardView.java
@@ -7,12 +7,16 @@ import inc.flide.vim8.MainInputMethodService;
 import inc.flide.vim8.R;
 import inc.flide.vim8.keyboardactionlisteners.MainKeypadActionListener;
 import inc.flide.vim8.preferences.SharedPreferenceHelper;
+import inc.flide.vim8.ui.Theme;
 import inc.flide.vim8.views.ConstraintLayoutWithSidebar;
 
 
 public class MainKeyboardView extends ConstraintLayoutWithSidebar {
 
     private MainKeypadActionListener actionListener;
+    private String prefSidebarLeftKey;
+    private SharedPreferenceHelper sharedPreferenceHelper;
+    private LayoutInflater inflater;
 
     public MainKeyboardView(Context context) {
         super(context);
@@ -31,29 +35,31 @@ public class MainKeyboardView extends ConstraintLayoutWithSidebar {
 
     public void initialize(Context context) {
         actionListener = new MainKeypadActionListener((MainInputMethodService) context, this);
-        setupMainKeyboardView(context);
+        inflater = (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+        prefSidebarLeftKey = context.getString(R.string.pref_sidebar_left_key);
+        Theme.getInstance(context).onChange(this::setColors);
+        sharedPreferenceHelper = SharedPreferenceHelper
+                .getInstance(context)
+                .addListener(this::initializeView, prefSidebarLeftKey);
+        initializeView();
+    }
+
+    private void initializeView() {
+        setupMainKeyboardView();
         setupButtonsOnSideBar(actionListener);
         setColors();
         setHapticFeedbackEnabled(true);
-
-        SharedPreferenceHelper.getInstance(context).addListener(this::setColors);
     }
 
-    private void setupMainKeyboardView(Context context) {
-        LayoutInflater inflater =
-                (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-
-        boolean preferredSidebarLeft = SharedPreferenceHelper
-                .getInstance(context)
-                .getBoolean(
-                        context.getString(R.string.pref_sidebar_left_key),
-                        true);
-
+    private void setupMainKeyboardView() {
+        boolean preferredSidebarLeft = sharedPreferenceHelper.getBoolean(prefSidebarLeftKey, true);
+        removeAllViews();
         if (preferredSidebarLeft) {
             inflater.inflate(R.layout.main_keyboard_left_sidebar_view, this, true);
         } else {
             inflater.inflate(R.layout.main_keyboard_right_sidebar_view, this, true);
         }
+
     }
