+++ 8vim/src/main/res/layout/x_keypad_layout.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<inc.flide.vim8.ui.views.mainkeyboard.XpadView xmlns:android="http://schemas.android.com/apk/res/android"
+<inc.flide.vim8.views.mainkeyboard.XpadView xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_gravity="bottom"
